+++ .changeset/quiet-birds-fly.md
@@ -1,7 +1,10 @@
 ---
 "@cloudflare/vite-plugin": patch
+"@cloudflare/workers-utils": patch
 ---
 
 Filter compatibility date fallback warning when no update is available
 
 The compatibility date warning from workerd (e.g., "The latest compatibility date supported by the installed Cloudflare Workers Runtime is...") is now only shown when a newer version of `@cloudflare/vite-plugin` is available. This matches the behavior in Wrangler and reduces noise when the user is already on the latest version.
+
+The update-check logic has been extracted to `@cloudflare/workers-utils` so it can be shared across packages.
