+++ packages/vite-plugin-cloudflare/src/update-check.ts
@@ -1,42 +1,32 @@
-import checkForUpdate from "update-check";
-import type { Result } from "update-check";
-
-/**
- * Checks if a newer version of the vite-plugin-cloudflare package is available.
- *
- * This function dynamically imports the package.json to get the current version,
- * then uses the `update-check` library to query npm for the latest version.
- * The dist tag used for comparison depends on the current version - "beta" for
- * pre-release versions (0.0.0-*) and "latest" for stable versions.
- *
- * @returns The latest available version string if an update is available, or `undefined` if the package is up-to-date or the check fails
- */
-async function doUpdateCheck(): Promise<string | undefined> {
-	let update: Result | null = null;
-	// Use dynamic import with JSON assertion to avoid bundler issues
-	const pkg = (
-		await import("../package.json", {
-			with: { type: "json" },
-		})
-	).default;
-	try {
-		// default cache for update check is 1 day
-		update = await checkForUpdate(
-			{ name: pkg.name, version: pkg.version },
-			{
-				distTag: pkg.version.startsWith("0.0.0") ? "beta" : "latest",
-			}
-		);
-	} catch {
-		// ignore error
-	}
-	return update?.latest;
-}
+import { doUpdateCheck } from "@cloudflare/workers-utils";
 
 // Memoise update check promise, so we can call this multiple times as required
 // without having to prop drill the result. It's unlikely to change through the
 // process lifetime.
 let updateCheckPromise: Promise<string | undefined>;
+
+/**
+ * Checks if a newer version of `@cloudflare/vite-plugin` is available on npm.
+ *
+ * The result is memoised so the check is only performed once per process
+ * lifetime — callers can invoke this freely without worrying about redundant
+ * network requests.
+ *
+ * @returns The latest available version string if an update exists, or `undefined` if up-to-date or the check fails
+ */
 export function updateCheck(): Promise<string | undefined> {
-	return (updateCheckPromise ??= doUpdateCheck());
+	return (updateCheckPromise ??= resolveUpdateCheck());
+}
+
+async function resolveUpdateCheck(): Promise<string | undefined> {
+	try {
+		const pkg = (
+			await import("../package.json", {
+				with: { type: "json" },
+			})
+		).default;
+		return doUpdateCheck(pkg.name, pkg.version);
+	} catch {
+		return undefined;
+	}
 }
