+++ packages/workers-utils/src/index.ts
@@ -110,3 +110,5 @@ export type { Counter } from "./prometheus-metrics";
 export type { Tunnel, TunnelOptions } from "./tunnel";
 export { startTunnel } from "./tunnel";
 export { spawnCloudflared } from "./cloudflared";
+
+export { doUpdateCheck } from "./update-check";
